[WIP] Add cluster_rule_helper
